requirements.txt
geopandas
rasterio
pillow
scikit-learn
tensorboard
tqdm
tifffile
timm
einops
opencv-python
rioxarray
gdown
ptflops
google-cloud-storage
omegaconf
pyDataverse
pytest
yacs
wandb
hydra-core

